+++ app/policies/activity_policy.rb
@@ -1,0 +1,5 @@
+class ActivityPolicy < ApplicationPolicy
+  def update?
+    (record.plan.user_id == user.id) || user.admin?
+  end
+end