+++ app/policies/time_range_policy.rb
@@ -1,0 +1,5 @@
+class TimeRangePolicy < ApplicationPolicy
+  def update?
+    (record.user_id == user.id) || user.admin?
+  end
+end