+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2020-12-15
+
+### Added
+
+- Add Pundit for authorization.
+- Authorize Plan and Signoff controllers.
+- Allow admins and signoffs to index plans.
+- Hide Plan and Signoff actions if not authorized.
+
 ## [0.6.2] - 2020-12-11
 
 ### Fixed